move remaining non component files to utils
